crash bug fixed, fixed warning;

+++ core/rectification/multicameraTriangulator.h
@@ -148,7 +148,7 @@ public:
     **/
    int getSize()
    {
-       return P.size();
+       return (int)P.size();
    }
 
    /**
